ref(iroh-net): Move PathState to its own module
There are no functional changes, this only moves code and fixes up
imports.
