+++ iroh-net/src/magicsock/node_map/udp_paths.rs
@@ -15,7 +15,8 @@ use tracing::warn;
 use crate::disco::SendAddr;
 
 use super::best_addr::{self, BestAddr};
-use super::node_state::{PathState, PongReply};
+use super::node_state::PongReply;
+use super::path_state::PathState;
 use super::IpPort;
 
 /// The address on which to send datagrams over UDP.
